Update format user record file to include sendSqsMessage function
